pyproject.toml
[tool.poetry]
name = "jocoding"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
python-dotenv = "^1.0.1"
langchain = "^0.3.13"
langchain-community = "^0.3.13"
langchain-openai = "^0.2.14"
openai = "^1.58.1"
anthropic = "^0.42.0"
google-generativeai = "^0.8.3"
langchain-google-genai = "^2.0.7"
streamlit = "^1.41.1"
pypdf = "^5.1.0"
chromadb = "^0.5.23"
tiktoken = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
